1. Add a minimal atomic example called example/atomic/get_started.cpp. 2. Advance version number in source to today.
optimize.hpp: Avoid warning with signed address type (used for testing).
atomic_base.hpp: Fix documentation, only zero order forward required. 
reciprocal.cpp: fix spelling of now.
reverse.hpp: Fix bug in checking for nan during reverse mode.

git-svn-id: https://projects.coin-or.org/svn/CppAD/trunk@2897 0e562018-b8fb-0310-a933-ecf073c7c197

+++ cppad/local/optimize.hpp
@@ -1641,7 +1641,7 @@ void optimize(
 			case UsravOp:
 			CPPAD_ASSERT_NARG_NRES(op, 1, 0);
 			new_arg[0] = tape[arg[0]].new_var;
-			if( new_arg[0] < num_var )
+			if( size_t(new_arg[0]) < num_var )
 			{	rec->PutArg(new_arg[0]);
 				rec->PutOp(UsravOp);
 			}
